Copy image from previous release
For some versions and architectures, which are not supported anymore,
upstream has shutdown the pkg site. For some of these architectures
there is no pkg files on the install media to bootstrap from. This
also means now that it's not possible to rebuild these images,
without removing these packages. We do this to avoid breaking our
users.