Improve checking if upstream pkg site is available
It's not enough that the actual package is available. There's a
corresponding signature file that needs to be present as well. The
`pkg.txz.sig` is not present anymore for x86-64, this causes FreeBSD
13.0 to fail to build. Seems to work fine for the other versions of
FreeBSD 13, they probably use `pkg.pkg`, which still has a signature
file `pkg.pkg.sig` available.